Added ItemList UI sounds logic
ItemLists are handy for simple inventories and so having the basics of selected, and activated emitting sounds seemed a good addition.

I've followed the same coding scheme you used for others, with a dedicated function to catch the unneeded passed index numbers and play the stream, similar to the line edit player, etc functions.